src/main/java/br/com/ras/java/leetcode/codechallenge/week2/DiameterBinaryTree.java
package br.com.ras.java.leetcode.codechallenge.week2;

import br.com.ras.java.leetcode.codechallenge.shared.BinaryTree;
import br.com.ras.java.leetcode.codechallenge.shared.TreeNode;

/**
 * Diameter of Binary Tree
 *
 * Given a binary tree, you need to compute the length of the diameter of the tree.
 * The diameter of a binary tree is the length of the longest path between any two nodes in a tree.
 * This path may or may not pass through the root.
 *
 * Example:
 * Given a binary tree
 *
 *           1
 *          / \
 *         2   3
 *        / \
 *       4   5
 *
 * Return 3, which is the length of the path [4,2,1,3] or [5,2,1,3].
 *
 * Note: The length of path between two nodes is represented by the number of edges between them.
 */
public class DiameterBinaryTree {

    public int solution(TreeNode root) {
        final BinaryTree tree = new BinaryTree(root);
        return tree.diameter();
    }

    int ans;
    public int best(TreeNode root) {
        ans = 1;
        depth(root);
        return ans - 1;
    }

    private int depth(TreeNode node) {
        if (node == null) return 0;
        int L = depth(node.left);
        int R = depth(node.right);
        ans = Math.max(ans, L+R+1);
        return Math.max(L, R) + 1;
    }
}
